Move IGVF backend query to models.py
